feat(profile): handle avatar upload via event bus
- Use asynchronous UploadAvatarEvent to decouple avatar upload from profile update.
