+++ backend/apps/profile/src/profile.module.ts
@@ -15,6 +15,7 @@ import { CqrsModule } from '@nestjs/cqrs';
 import { GetUserProfileHandler } from './queries/get-user-profile/get-user-profile.handler';
 import { GetAllUserProfilesHandler } from './queries/get-all-user-profiles/get-all-user-profiles.handler';
 import { UpdateProfileHandler } from './commands/update-profile/update-profile.handler';
+import { UploadAvatarEventHandler } from './events/upload-avatar.handler';
 
 dotenv.config();
 
@@ -45,6 +46,7 @@ dotenv.config();
     GetUserProfileHandler,
     GetAllUserProfilesHandler,
     UpdateProfileHandler,
+    UploadAvatarEventHandler,
   ],
 })
 export class ProfileModule {}
