+++ backend/libs/common/src/events/index.ts
@@ -1,0 +1,1 @@
+export * from './profile';
