+++ backend/libs/common/src/events/profile/index.ts
@@ -1,0 +1,1 @@
+export * from './upload-avatar.event';
