+++ backend/libs/common/src/events/profile/upload-avatar.event.ts
@@ -1,0 +1,6 @@
+export class UploadAvatarEvent {
+  constructor(
+    public readonly userId: number,
+    public readonly avatarPayload: any,
+  ) {}
+}
