+++ backend/libs/common/src/index.ts
@@ -1,3 +1,4 @@
 export * from './dto';
 export * from './constants';
 export * from './redis';
+export * from './events';
